API Now Working VIA POST
When we get to JSONArray:
If it's a string instead of a JSONArray, it's an error of some sort.
If it's null, there is no results.
If it's a Boolean, it's a pass or fail.

We can try to check the type before we we do the Try/Catch block around
the JSONArray looping.

WE DID IT! Not much left.

We're left with JSONObjects, which can be parsed and inserted into our
local database, deleted from the remote database, and migrated/synced
back into the remote tasks table.

:)